requirements.txt
numpy
scipy
pytest
hatch
pyyaml
matplotlib
pandas